Version 1.32 Added 3 File Modified 23 Files Added Screen Scaling Made the app look alot nicer and cleaner.

+++ app/src/main/java/delware/apps/techsupport_scampermobile/RunLog.java
@@ -3,7 +3,7 @@ package delware.apps.techsupport_scampermobile;
 public class RunLog {
         public float Distance;
         public int Hours;
-        public float Minutes;
+        public int Minutes;
         public int calories;
         public float speed;
         public String date;
@@ -11,7 +11,7 @@ public class RunLog {
         public int UserId;
 
 
-    public RunLog(float distance, int hours, float minutes, int calories, String date, String cardioType) {
+    public RunLog(float distance, int hours, int minutes, int calories, String date, String cardioType) {
         Distance = distance;
         Hours = hours;
         Minutes = minutes;
@@ -37,11 +37,11 @@ public class RunLog {
             Hours = hours;
         }
 
-        public float getMinutes() {
+        public int getMinutes() {
             return Minutes;
         }
 
-        public void setMinutes(float minutes) {
+        public void setMinutes(int minutes) {
             Minutes = minutes;
         }
